+++ requirements.txt
@@ -1,6 +1,8 @@
 beautifulsoup4==4.4.1
-click==6.2
+click==6.6
+coverage==4.0.3
 FormEncode==1.3.0
-lxml==3.5.0
+lxml==3.6.0
+py==1.4.31
+pytest==2.9.1
 PyYAML==3.11
-wheel==0.29.0
